requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.0.2
Flask-WTF==1.0.0
graphviz==0.18.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.0.1
mysqlclient==2.1.0
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.4
passlib==1.7.4
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
urllib3==1.26.7
uvicorn==0.15.0
Werkzeug==2.0.2
WTForms==3.0.0
